requirements.txt
jsonschema
haystack-ai
pygments
transformers[agents]
openai
asyncio
python-dotenv
llama-index
traceback2
pydantic
GitPython